Use the js file from rawgithub.com instead of raw.github.com.
Github serves the file from raw.github.com with the plain/text MIME type. Google Chrome (and maybe others) complains about this and does not execute the javascript.

On the other side, rawgithub.com is a service created specifically to solve this problem. It takes the file from raw.github.com and serves it with the correct MIME type.